+++ CHANGELOG.md
@@ -12,6 +12,8 @@ All changes to the **Hearts of Modding** extension will be documented in this fi
 
 - **Unit (`common/units/*.txt`) scanner.** Units are validated within OOBs.
 
+- **Missing bookmark related keywords.**
+
 - `x` and `y` (in `.gui`s, OOBs, etc) are now semantically highlighted in a distinctive color.
 
 ### 🩹 Fixed
@@ -19,6 +21,8 @@ All changes to the **Hearts of Modding** extension will be documented in this fi
 - **Fixed an air wings OOB edge-case where HoM was emitting a `duplicate_key` warning for multiple `name` keys being under the same state scope under respective equipment types (planes).** This is *required* to name air wings when there are more than one of them in the same state/airfield.
 
 - **Fixed an issue where `did_change_watched_files` only updated the scan data, never re-validated open documents whose diagnostics could depend on that data.**
+
+- **Fixed semantic highlighting of the date in a bookmark.**
 
 ## [v0.17.0] - 2026-06-07